requirements.txt
# StudyMate AI - Complete Requirements
# Core Backend
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# AI & ML
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.21.0
faiss-cpu>=1.7.0

# File Processing
PyMuPDF>=1.23.0
python-pptx>=0.6.21
python-docx>=1.0.0
docx2txt>=0.8

# Voice Processing
openai-whisper>=20231117
elevenlabs>=0.2.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
requests>=2.31.0
aiofiles>=23.0.0
jinja2>=3.1.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
markdown>=3.5.1
html2text>=2024.2.26
reportlab>=4.0.0
weasyprint>=60.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0